change js style to ES Module
